coolsite/requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==3.2.15
django-filter==21.1
djangorestframework==3.13.1
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.30
Pygments==2.13.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing-extensions==4.1.1
wcwidth==0.2.5
